requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
requests==2.31.0
redis==5.0.1
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2
xgboost==2.0.2
joblib==1.3.2
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pydantic==2.11.7
pydantic-settings==2.10.1
psycopg==3.2.9
typer==0.9.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
python-multipart==0.0.6
seaborn==0.12.2